feat: système XP/progression/maîtrise (spec §1–5)
- XP par question : (2 + 8×(1−maîtrise_avant)) × facteur (1.0/0.75/0.0)
- Maîtrise pondérée : historique flottant 0.0/0.75/1.0 (DB v14, reset stats)
- Niveau calculé depuis xp_total, affiché avec barre de progression dans Réglages
- Snapshot maitriseAvant passé au Quiz depuis lancerQuiz()
- Bonus Double XP appliqué sur resultat.xpQuiz dans quiz_screen
- NiveauHelper : niveauDepuisXp, xpCumulePourNiveau, progressionNiveau

+++ mobile/lib/services/adaptive_selector.dart
@@ -5,7 +5,9 @@ import '../models/question.dart';
 /// Statistiques de maitrise d'une question (fenetre glissante de 10 tentatives).
 class QuestionStats {
   final int nbAffichee;
-  final int nbCorrecte;
+  // Somme ponderee des valeurs de reussite (1.0 / 0.75 / 0.0).
+  // Peut etre fractionnaire (ex. 7.5 sur 10 tentatives).
+  final double nbCorrecte;
   final DateTime? derniereTentative;
 
   const QuestionStats({
